Refact : 베포전 상수 데이터 정리
