+++ package.json
@@ -18,6 +18,7 @@
 		"axios": "^1.4.0",
 		"chart.js": "^4.3.0",
 		"framer-motion": "^10.12.21",
+		"history": "^5.3.0",
 		"react": "^18.2.0",
 		"react-chartjs-2": "^5.2.0",
 		"react-datepicker": "^4.14.0",
