+++ src/components/layout/Sidebar.tsx
@@ -7,7 +7,7 @@ import { CSSTransition } from "react-transition-group";
 import "../../styles/Sidebar.css";
 import { useContext } from "react";
 import { AuthStateContext } from "../../context/AuthStateContext";
-import { LoudOli, OliImage } from "../../utils/constant/contants";
+import { LoudOli, OliImage } from "../../utils/constant/image";
 
 type Props = {
 	isNameExisted: boolean;
