+++ src/utils/common/history.ts
@@ -1,0 +1,4 @@
+import { createBrowserHistory } from "history";
+
+const history = createBrowserHistory();
+export default history;
