+++ src/utils/constant/image.ts
@@ -1,0 +1,14 @@
+// 이미지
+import hero from "../../assets/logo/hero.jpeg";
+import logo from "../../assets/logo/logo.png";
+import oli from "../../assets/logo/oli.png";
+import oliBody from "../../assets/logo/oliBody.png";
+import loudOli from "../../assets/logo/louderOli.png";
+import louderOli from "../../assets/logo/loudOli.png";
+
+export const heroImage = hero;
+export const LogoImage = logo;
+export const OliImage = oli;
+export const OliBodyImage = oliBody;
+export const LoudOli = loudOli;
+export const LouderOli = louderOli;
