+++ src/utils/constant/target.ts
@@ -1,0 +1,17 @@
+export const GOAL_TITLE = "당신의 목표는 무엇인가요?";
+export const GOAL_DESCRIPTION = "이루고자 하는 최종 목표를 적어주세요";
+
+export const SUBGOAL_TITLE = "목표를 위한 세분화 목표와 루틴을 작성해주세요";
+export const SUBGOAL_DESCRIPTION =
+	"목표를 달성하기 위해서는 반드시 필요한 작업들이 있어요. 바로 이게 그 작업이에요. 작은 목표를 세우고 루틴을 만든다면 목표를 달성하는데 도움이 될거예요 파이팅!";
+
+export const DURATION_TITLE = "언제까지 목표를 달성하기 위해 노력하실건가요?";
+export const DURATION_DESCRIPTION =
+	"마감일은 언제나 우리에게 긴장감을 주죠. 하지만 너무 부담 갖지 않아도 돼요";
+
+// export const PENALTY_TITLE = "혹시나 ";
+// export const PENALTY_DESCRIPTION = "대목표를 작성해주시면 좋아요.";
+
+export const LAST_TITLE = "최종 확인";
+export const LAST_DESCRIPTION =
+	"목표를 달성하기 위해 하위 목표 그리고 루틴을 세웠어요. 아래의 정보를 확인하시고 변경하고자 하시면 이전 버튼으로 누르시면 돼요";
